--- src/constants.js
@@ -1,4 +1,0 @@
-export const headingText = {
-    LISTHEADER: 'To Do Lists',
-    TASKHEADER: 'Tasks'
-}
